+++ src/ls_orders.c
@@ -12,58 +12,7 @@
 
 #include "ft_ls.h"
 
-void	order_list_fill_stat(t_ls_order *order_list)
-{
-	t_passwd	*passwd;
-	t_group		*group;
-	t_entry		*entry;
-
-	while (order_list != NULL)
-	{
-		entry = order_list->list;
-		while (entry != NULL)
-		{
-			ls_nullptr((entry->full_name = ft_strjoin_3(order_list->name, "/", entry->name)));
-			if (stat(entry->full_name, &(entry->stat)) != 0)
-				ls_unknown_error(errno);
-
-			if (entry->dirent.d_type == DT_LNK)
-			{
-				//TODO delete me
-				ft_printf("found link\n");
-				ft_printf("full name is %s\n", entry->full_name);
-				char *full_buf = ft_memalloc(entry->stat.st_size);
-				ssize_t size = readlink(entry->full_name, full_buf, entry->stat.st_size + 1);
-				size++;
-				//
-//			if (size < 0)
-//			{
-//				ft_printf("meh, error\n");
-//			}
-//			else
-//			{
-				char *new_buf = ft_strdup(full_buf);
-				ft_free(full_buf);
-				ft_printf("look, (%s)!\n", new_buf);
-//			}
-			}
-
-			if ((passwd = getpwuid(entry->stat.st_uid)) != NULL)
-				entry->owner = ft_strdup(passwd->pw_name);
-			else
-				ls_unknown_error(errno);
-			if ((group = getgrgid(entry->stat.st_gid)) != NULL)
-				entry->group = ft_strdup(group->gr_name);
-			else
-				ls_unknown_error(errno);
-			entry = entry->entry_next;
-		}
-		order_list = order_list->next;
-	}
-}
-
-static
-t_ls_order		*ls_order_malloc(char *order_name)
+t_ls_order	*ls_order_malloc(char *order_name)
 {
 	t_ls_order		*order;
 
@@ -78,8 +27,7 @@ t_ls_order		*ls_order_malloc(char *order_name)
 	return (order);
 }
 
-static
-t_ls_order		*ls_order_create(t_input *input, char *order_name)
+t_ls_order	*ls_order_create(t_input *input, char *order_name)
 {
 	t_ls_order		*order;
 
@@ -106,8 +54,7 @@ t_ls_order		*ls_order_create(t_input *input, char *order_name)
 	return (order);
 }
 
-static
-t_ls_order		*ls_order_create_rec(t_input *input, char *order_name)
+t_ls_order	*ls_order_create_rec(t_input *input, char *order_name)
 {
 	t_entry			*entry;
 	t_ls_order		*order;
@@ -120,26 +67,25 @@ t_ls_order		*ls_order_create_rec(t_input *input, char *order_name)
 	entry = order->list;
 	while (entry != NULL)
 	{
-		if (ft_strcmp(entry->name, ".") != 0 && ft_strcmp(entry->name, "..") != 0)
+		if (ft_strcmp(entry->name, ".") != 0
+					&& ft_strcmp(entry->name, "..") != 0)
 		{
 			order->next = ls_order_create_rec(input,
 				ft_strjoin_3(order_name, "/", entry->name));
 			while (order->next != NULL)
 				order = order->next;
 		}
-		//
 		entry = entry->entry_next;
 	}
 	return (order_rec);
 }
 
-static
-t_ls_order			*ls_order_list_create_rec(t_input *input,
+t_ls_order	*ls_order_list_create_rec(t_input *input,
 					t_ls_order *order_list)
 {
 	t_ls_order			*order_tmp;
 	t_ls_order			*order;
-	size_t 				i;
+	size_t				i;
 
 	ls_nullptr(input);
 	order_tmp = NULL;
@@ -162,13 +108,12 @@ t_ls_order			*ls_order_list_create_rec(t_input *input,
 	return (order_list);
 }
 
-static
-t_ls_order			*ls_order_list_create_plain(t_input *input,
+t_ls_order	*ls_order_list_create_plain(t_input *input,
 					t_ls_order *order_list)
 {
-	t_ls_order			*order_tmp;
-	t_ls_order			*order;
-	size_t 				i;
+	t_ls_order		*order_tmp;
+	t_ls_order		*order;
+	size_t			i;
 
 	ls_nullptr(input);
 	order_tmp = NULL;
@@ -187,14 +132,4 @@ t_ls_order			*ls_order_list_create_plain(t_input *input,
 		i++;
 	}
 	return (order_list);
-}
-
-t_ls_order			*ls_order_list_create(t_input *input)
-{
-	t_ls_order			*order_list;
-
-	order_list = NULL;
-	if (input->rec == TRUE)
-		return (ls_order_list_create_rec(input, order_list));
-	return (ls_order_list_create_plain(input, order_list));
 }
